fix: delete/reactive client use case & vehicle category

+++ LavaRapido/src/main/java/com/example/lavarapido/usecases/Client/DeleteClientUseCase.java
@@ -8,20 +8,11 @@ import java.util.Optional;
 
 public class DeleteClientUseCase {
 
-    private ClientDAO clientDAO;
+    private final ClientDAO clientDAO;
 
     public DeleteClientUseCase(ClientDAO clientDAO) {
         this.clientDAO = clientDAO;
     }
-
-    //IMPLEMENTAÇÃO INCOMPLETA POR CAUSA DOS AGENDAMENTOS
-
-    /*
-    * Para realizar a remoçao / inativação do cliente, precisamos verificar seus Agendamentos (Appointments)
-    * Esses 'Appointments' é apontado como uma classe nos diagramas de sequencia, mas nao é indicado essa
-    * classe.
-    * É necessário a criação dessa classe? Como podemos implementar essa parte?
-    */
 
     public Optional<List<Client>> findClientsByName(String name) {
         return clientDAO.findByName(name);
@@ -31,13 +22,11 @@ public class DeleteClientUseCase {
         if (client == null || clientDAO.findOne(client.getCpfString()).isEmpty())
             throw new EntityNotFoundException("Client not found.");
 
-        /*
-        * if (clientDAO.getAppointments(client).isEmpty())
-        *   return clientDAO.delete(client)
-        *
-        * return clientDAO.inactivate(client) --> update?
-        * */
-        return clientDAO.delete(client);
+
+        if (client.getSchedulings().isEmpty())
+           return clientDAO.delete(client);
+
+        return clientDAO.update(client);
     }
 
 }
